Version 2.1-BETA

+++ src/sys/boot/ficl/tools.c
@@ -54,7 +54,7 @@
 ** Specify breakpoint default action
 */
 
-/* $FreeBSD: release/10.0.0/sys/boot/ficl/tools.c 249223 2013-04-07 05:53:42Z kientzle $ */
+/* $FreeBSD: stable/10/sys/boot/ficl/tools.c 271135 2014-09-04 21:01:10Z emaste $ */
 
 #ifdef TESTMAIN
 #include <stdlib.h>
@@ -201,7 +201,7 @@ static void seeColon(FICL_VM *pVM, CELL *pc)
 			*cp++ = '>';
 		else
 			*cp++ = ' ';
-        cp += sprintf(cp, "%3d   ", pc-param0);
+        cp += sprintf(cp, "%3d   ", (int)(pc-param0));
         
         if (isAFiclWord(pd, pFW))
         {
@@ -239,40 +239,40 @@ static void seeColon(FICL_VM *pVM, CELL *pc)
             case IF:
                 c = *++pc;
                 if (c.i > 0)
-                    sprintf(cp, "if / while (branch %d)", pc+c.i-param0);
+                    sprintf(cp, "if / while (branch %d)", (int)(pc+c.i-param0));
                 else
-                    sprintf(cp, "until (branch %d)",      pc+c.i-param0);
+                    sprintf(cp, "until (branch %d)",      (int)(pc+c.i-param0));
                 break;                                                           
             case BRANCH:
                 c = *++pc;
                 if (c.i == 0)
-                    sprintf(cp, "repeat (branch %d)",     pc+c.i-param0);
+                    sprintf(cp, "repeat (branch %d)",     (int)(pc+c.i-param0));
                 else if (c.i == 1)
-                    sprintf(cp, "else (branch %d)",       pc+c.i-param0);
+                    sprintf(cp, "else (branch %d)",       (int)(pc+c.i-param0));
                 else
-                    sprintf(cp, "endof (branch %d)",       pc+c.i-param0);
+                    sprintf(cp, "endof (branch %d)",      (int)(pc+c.i-param0));
                 break;
 
             case OF:
                 c = *++pc;
-                sprintf(cp, "of (branch %d)",       pc+c.i-param0);
+                sprintf(cp, "of (branch %d)",       (int)(pc+c.i-param0));
                 break;
 
             case QDO:
                 c = *++pc;
-                sprintf(cp, "?do (leave %d)",  (CELL *)c.p-param0);
+                sprintf(cp, "?do (leave %d)",  (int)((CELL *)c.p-param0));
                 break;
             case DO:
                 c = *++pc;
-                sprintf(cp, "do (leave %d)", (CELL *)c.p-param0);
+                sprintf(cp, "do (leave %d)", (int)((CELL *)c.p-param0));
                 break;
             case LOOP:
                 c = *++pc;
-                sprintf(cp, "loop (branch %d)", pc+c.i-param0);
+                sprintf(cp, "loop (branch %d)", (int)(pc+c.i-param0));
                 break;
             case PLOOP:
                 c = *++pc;
-                sprintf(cp, "+loop (branch %d)", pc+c.i-param0);
+                sprintf(cp, "+loop (branch %d)", (int)(pc+c.i-param0));
                 break;
             default:
                 sprintf(cp, "%.*s", pFW->nName, pFW->name);
